% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{estimate}
\alias{estimate}
\title{Estimate means and variances of many populations from samples of each}
\usage{
estimate(x_ij)
}
\arguments{
\item{x_ij}{list of numeric vectors, corresponding to independent random samples each from a different population}
}
\value{
data.frame with one row for each element of \code{x_ij} and columns corresponding to statistics of \code{x_ij} including the proposed estimator for the means, \code{t_i}, and for the variances, \code{w_i}.
}
\description{
Implements the general case of the non-parametric linear empirical Bayes
estimator of many means and many variances as described in the same named
publication by Herbert Robbins (1985) <doi:10.1073/pnas.82.6.157>.
}
\examples{
estimate(x_ij = split(iris$Sepal.Length, f = iris$Species))

}
